updating code

+++ code/metrics.py
@@ -14,22 +14,29 @@ from sklearn.metrics import average_precision_score
 ## We can see how frequently the predictions from our algorithms coincide with the actual species. 
 ####################
 
-def metrics ():
-    print("Calculating metrics for " + sys.argv[1] + sys.argv[2] + " dataset\n")
-    resultsFile = 'results/' + sys.argv[1] + sys.argv[2] + '_3.tsv' #the file created from classify.py or from ensemblePredictions.py
-    originalDataFile = 'data/' + sys.argv[1] + 'Modified.csv'
-    outFile = 'results/' + sys.argv[1] + sys.argv[2] + 'Metrics_3.tsv'
+
+resultsFile = 'results/Classifications.tsv'#sys.argv[1] + sys.argv[2] + '_3.tsv' #the file created from classify.py or from ensemblePredictions.py
+outFile = 'results/Metrics.tsv'
+dataResults = pd.read_csv(resultsFile, sep = '\t')
+
+def metrics (dataName):
+    print("Calculating metrics for " + dataName + " dataset\n")
+    #resultsFile = 'results/Classifications.tsv'#sys.argv[1] + sys.argv[2] + '_3.tsv' #the file created from classify.py or from ensemblePredictions.py
+    originalDataFile = 'data/' + dataName + 'Modified.csv'
+    #outFile = 'results/' + sys.argv[1] + sys.argv[2] + 'Metrics_3.tsv'
     
-    dataResults = pd.read_csv(resultsFile, sep = '\t')
+    #dataResults = pd.read_csv(resultsFile, sep = '\t')
     ogData = pd.read_csv(originalDataFile)
 
     listOfClassifiers = dataResults["Classifier"].unique()
 
     numIterations = dataResults["Iteration"].max() + 1
-
-    with open(outFile, "w") as tsvFile:
-        tsvFile.write("DataName\tClassifier\tIteration\tPredictionType\tAccuracy\tf1_score\tf1_weighted\taverage_precision\troc_auc\n")
-
+    
+    if not os.path.exists (outFile):
+        with open(outFile, "w") as tsvFile:
+            tsvFile.write("DataName\tClassifier\tIteration\tPredictionType\tAccuracy\tf1_score\tf1_weighted\taverage_precision\troc_auc\n")
+    
+    with open(outFile, "a") as tsvFile:
         for classifier in listOfClassifiers:
             for iteration in range(1, numIterations):
                 y_pred = []
@@ -53,6 +60,8 @@ def metrics ():
                 
                 ### writing the information to the tsv file
                 print("Writing the metrics for " + classifier + " for iteration " + str(iteration))
-                tsvFile.write('\t'.join([str(sys.argv[1]), classifier, str(iteration), str(predictionType), str(accuracy), str(f1Score), str(f1Weighted), str(averagePrecision), str(rocAucScore)]) + '\n')
+                tsvFile.append('\t'.join([dataName, classifier, str(iteration), str(predictionType), str(accuracy), str(f1Score), str(f1Weighted), str(averagePrecision), str(rocAucScore)]) + '\n')
 
-metrics()
+
+for dataName in dataResults["DataName"].unique():
+    metrics(dataName)
